fix segfault in critbit string test
closes #3

when compiled with heavy optimizations, namely
gcc -O3 -flto -fwhole-program critbit.c critbit-test.c -o critbit-test

the "critbit" test segfaults, because the loop on line 360:

        for (cnt = 0; test_data[cnt]; ) {
                cnt++;
        }

assumes that the last element of test data is a NULL pointer.
with those optimizations set, this is not the case and cnt increases
to > 2000 elements, pointing to junk.